Feat/clean

+++ src/pages/api/like/post.ts
@@ -26,6 +26,11 @@ const handler: NextApiHandler<
     if (req.method === "POST") {
       const { postIdx } = req.body as ApiUploadLikeOfPostRequest;
 
+      // 존재하지 않는 게시글에 좋아요 추가 요청
+      const exPost = await prisma.post.findUnique({ where: { idx: +postIdx } });
+      if (!exPost)
+        return res.status(404).json({ message: "존재하지 않는 게시글입니다." });
+
       await prisma.postLike.create({
         data: {
           postLiker: { connect: { idx: req.user.idx } },
@@ -42,6 +47,11 @@ const handler: NextApiHandler<
     // 게시글 좋아요 삭제 요청
     if (req.method === "DELETE") {
       const postIdx = +req.query.postIdx!;
+
+      // 존재하지 않는 게시글에 좋아요 제거 요청
+      const exPost = await prisma.post.findUnique({ where: { idx: +postIdx } });
+      if (!exPost)
+        return res.status(404).json({ message: "존재하지 않는 게시글입니다." });
 
       await prisma.postLike.delete({
         where: {
